+++ GHEtool/test/unit-tests/test_main_class.py
@@ -7,8 +7,9 @@ import numpy as np
 import pygfunction as gt
 import pytest
 
-from GHEtool import GroundConstantTemperature, GroundFluxTemperature, FluidData, DoubleUTube, Borefield, \
-    CalculationSetup, FOLDER, MultipleUTube, EERCombined, ConstantFlowRate, TemperatureDependentFluidData
+from GHEtool import GroundConstantTemperature, GroundFluxTemperature, DoubleUTube, Borefield, \
+    CalculationSetup, FOLDER, MultipleUTube, EERCombined, ConstantFlowRate, TemperatureDependentFluidData, \
+    ConstantFluidData
 from GHEtool.Validation.cases import load_case
 from GHEtool.VariableClasses.LoadData import MonthlyGeothermalLoadAbsolute, HourlyGeothermalLoad, HourlyBuildingLoad, \
     HourlyBuildingLoadMultiYear, MonthlyBuildingLoadAbsolute
@@ -19,7 +20,8 @@ from GHEtool import CustomGFunction
 data = GroundConstantTemperature(3, 10)
 ground_data_constant = data
 data_ground_flux = GroundFluxTemperature(3, 10)
-fluidData = FluidData(0.2, 0.568, 998, 4180, 1e-3)
+fluidData = ConstantFluidData(0.568, 998, 4180, 1e-3)
+constantFlowData = ConstantFlowRate(mfr=0.2)
 pipeData = DoubleUTube(1, 0.015, 0.02, 0.4, 0.05)
 flowData = ConstantFlowRate(vfr=0.2)
 
@@ -61,7 +63,7 @@ def test_nb_of_boreholes():
     borefield = Borefield()
     assert borefield.number_of_boreholes == 0
     borefield = Borefield(borefield=copy.deepcopy(borefield_gt))
-    borefield.set_ground_parameters(data_ground_flux)
+    borefield.ground_data = data_ground_flux
     assert borefield.number_of_boreholes == 120
     borefield.set_borefield(gt.borefield.Borefield.rectangle_field(5, 5, 6, 6, 110, 0.1, 0.07))
     assert np.isclose(borefield.avg_tilt, 0)
@@ -268,7 +270,7 @@ def test_ground_data_custom_gfunction():
     assert not borefield.custom_gfunction is None
 
     # test for set function
-    borefield.set_ground_parameters(data_ground_flux)
+    borefield.ground_data = data_ground_flux
     assert borefield.ground_data == data_ground_flux
     assert borefield._ground_data == data_ground_flux
     assert borefield.custom_gfunction is None
@@ -294,7 +296,7 @@ def test_ground_data_jit_gfunction():
     assert np.any(borefield.gfunction_calculation_object.borehole_length_array)
 
     # test for set function
-    borefield.set_ground_parameters(data_ground_flux)
+    borefield.ground_data = data_ground_flux
     assert borefield.ground_data == data_ground_flux
     assert borefield._ground_data == data_ground_flux
     assert not np.any(borefield.gfunction_calculation_object.borehole_length_array)
@@ -304,8 +306,7 @@ def test_set_fluid_params():
     borefield = Borefield()
     assert borefield.borehole.fluid_data is None
     borefield.fluid_data = fluidData
-    assert borefield.borehole.fluid_data == fluidData.fluid_data
-    assert borefield.fluid_data == fluidData.fluid_data
+    assert borefield.fluid_data == fluidData
 
 
 def test_set_flow_params():
@@ -366,9 +367,10 @@ def test_Ahmadfard(ground_data, constant_Rb, result):
     borefield = Borefield()
     borefield.borefield = copy.deepcopy(borefield_gt)
     load = MonthlyGeothermalLoadAbsolute(*load_case(3))
-    borefield.set_ground_parameters(ground_data)
-    borefield.set_fluid_parameters(fluidData)
-    borefield.set_pipe_parameters(pipeData)
+    borefield.ground_data = ground_data
+    borefield.fluid_data = fluidData
+    borefield.flow_data = constantFlowData
+    borefield.pipe_data = pipeData
     borefield.calculation_setup(use_constant_Rb=constant_Rb)
     th, qh, qm, qa = load._calculate_last_year_params(True)
     assert np.isclose(borefield._Ahmadfard(th, qh, qm, qa, 0), result)
@@ -388,9 +390,10 @@ def test_Carcel(ground_data, constant_Rb, result):
     borefield.borefield = copy.deepcopy(borefield_gt)
     load = MonthlyGeothermalLoadAbsolute(*load_case(3))
 
-    borefield.set_ground_parameters(ground_data)
-    borefield.set_fluid_parameters(fluidData)
-    borefield.set_pipe_parameters(pipeData)
+    borefield.ground_data = ground_data
+    borefield.fluid_data = fluidData
+    borefield.flow_data = constantFlowData
+    borefield.pipe_data = pipeData
     borefield.calculation_setup(use_constant_Rb=constant_Rb)
     th, _, tcm, qh, qpm, qm = load._calculate_first_year_params(True)
     assert np.isclose(borefield._Carcel(th, tcm, qh, qpm, qm, 0), result)
@@ -585,6 +588,7 @@ def test_reynolds_number():
     borefield = Borefield()
     borefield.pipe_data = pipeData
     borefield.fluid_data = fluidData
+    borefield.flow_data = constantFlowData
     assert np.isclose(4244.131815783876, borefield.Re)
 
 
@@ -834,7 +838,7 @@ def test_set_options_gfunction_calculation():
 
 def test_gfunction():
     borefield = Borefield()
-    borefield.set_ground_parameters(data_ground_flux)
+    borefield.ground_data = data_ground_flux
     borefield.create_rectangular_borefield(10, 10, 6, 6, 100, 1, 0.075)
     borefield.H = 100_000
     with pytest.raises(UnsolvableDueToTemperatureGradient):
@@ -853,7 +857,7 @@ def test_gfunction():
 
 def test_gfunction_with_irregular_depth():
     borefield = Borefield()
-    borefield.set_ground_parameters(data_ground_flux)
+    borefield.ground_data = data_ground_flux
     temp = [
         gt.boreholes.Borehole(100, 4, 0.075, 0, 0),
         gt.boreholes.Borehole(150, 4, 0.075, 10, 0),
@@ -866,7 +870,7 @@ def test_gfunction_with_irregular_depth():
     assert not np.array_equal(borefield.gfunction([6000, 60000, 600000]), g_values)
 
     borefield = Borefield()
-    borefield.set_ground_parameters(data_ground_flux)
+    borefield.ground_data = data_ground_flux
     temp = [
         gt.boreholes.Borehole(100, 4, 0.075, 0, 0),
         gt.boreholes.Borehole(150, 4, 0.075, 10, 0),
@@ -886,7 +890,7 @@ def test_gfunction_with_irregular_depth():
 def test_load_duration(monkeypatch):
     borefield = Borefield()
     monkeypatch.setattr(plt, "show", lambda: None)
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     borefield.borefield = copy.deepcopy(borefield_gt)
     load = HourlyBuildingLoad(efficiency_heating=10 ** 6, efficiency_cooling=10 * 66)
     load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
@@ -898,7 +902,7 @@ def test_load_duration(monkeypatch):
 def test_optimise_load_profile_power(monkeypatch):
     borefield = Borefield()
     monkeypatch.setattr(plt, "show", lambda: None)
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     borefield.borefield = copy.deepcopy(borefield_gt)
     load = HourlyBuildingLoad(efficiency_heating=10 ** 6, efficiency_cooling=10 * 66)
     load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
@@ -913,7 +917,7 @@ def test_optimise_load_profile_power_multiyear(monkeypatch):
     # multiyear should also have a multiyear as output
     borefield = Borefield()
     monkeypatch.setattr(plt, "show", lambda: None)
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     borefield.borefield = copy.deepcopy(borefield_gt)
     load = HourlyBuildingLoad(efficiency_heating=10 ** 6, efficiency_cooling=10 * 66)
     load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
@@ -929,7 +933,7 @@ def test_optimise_load_profile_power_multiyear(monkeypatch):
 def test_optimise_load_profile_energy(monkeypatch):
     borefield = Borefield()
     monkeypatch.setattr(plt, "show", lambda: None)
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     borefield.borefield = copy.deepcopy(borefield_gt)
     load = HourlyBuildingLoad(efficiency_heating=10 ** 6, efficiency_cooling=10 * 66)
     load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
@@ -943,7 +947,7 @@ def test_optimise_load_profile_energy(monkeypatch):
 def test_optimise_borefield_small_power(monkeypatch):
     borefield = Borefield()
     monkeypatch.setattr(plt, "show", lambda: None)
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     borefield.create_rectangular_borefield(5, 1, 6, 6, 100)
     load = HourlyBuildingLoad(efficiency_heating=10 ** 6, efficiency_cooling=10 * 66)
     load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
@@ -956,7 +960,7 @@ def test_optimise_borefield_small_power(monkeypatch):
 def test_optimise_borefield_small_energy(monkeypatch):
     borefield = Borefield()
     monkeypatch.setattr(plt, "show", lambda: None)
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     borefield.create_rectangular_borefield(5, 1, 6, 6, 100)
     load = HourlyBuildingLoad(efficiency_heating=10 ** 6, efficiency_cooling=10 * 66)
     load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
@@ -968,7 +972,7 @@ def test_optimise_borefield_small_energy(monkeypatch):
 def test_optimise_borefield_wrong_threshold_power(monkeypatch):
     borefield = Borefield()
     monkeypatch.setattr(plt, "show", lambda: None)
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     borefield.create_rectangular_borefield(5, 1, 6, 6, 100)
     load = HourlyBuildingLoad(efficiency_heating=10 ** 6, efficiency_cooling=10 * 66)
     load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
@@ -980,7 +984,7 @@ def test_optimise_borefield_wrong_threshold_power(monkeypatch):
 def test_optimise_borefield_wrong_threshold_energy(monkeypatch):
     borefield = Borefield()
     monkeypatch.setattr(plt, "show", lambda: None)
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     borefield.create_rectangular_borefield(5, 1, 6, 6, 100)
     load = HourlyBuildingLoad(efficiency_heating=10 ** 6, efficiency_cooling=10 * 66)
     load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
@@ -994,7 +998,7 @@ def test_calculate_quadrants_without_data():
     borefield.borefield = copy.deepcopy(borefield_gt)
     borefield.set_max_avg_fluid_temperature(18)
     borefield.load = MonthlyGeothermalLoadAbsolute(*load_case(2))
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     borefield.calculate_quadrant()
 
 
@@ -1034,7 +1038,7 @@ def test_load_load():
 
 def test_calculate_temperature_profile():
     borefield = Borefield()
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     load = MonthlyGeothermalLoadAbsolute(*load_case(1))
     borefield.load = load
     with pytest.raises(ValueError):
@@ -1047,7 +1051,7 @@ def test_optimise_load_profile_power_without_hourly_data():
     with pytest.raises(ValueError):
         optimise_load_profile_power(borefield, borefield.load)
     borefield.load = HourlyBuildingLoad()
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     borefield.load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
     borefield.create_rectangular_borefield(10, 10, 6, 6, 150)
     optimise_load_profile_power(borefield, borefield.load)
@@ -1059,7 +1063,7 @@ def test_optimise_load_profile_energy_without_hourly_data():
     with pytest.raises(ValueError):
         optimise_load_profile_energy(borefield, borefield.load)
     borefield.load = HourlyBuildingLoad()
-    borefield.set_ground_parameters(ground_data_constant)
+    borefield.ground_data = ground_data_constant
     borefield.load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
     borefield.create_rectangular_borefield(10, 10, 6, 6, 150)
     optimise_load_profile_energy(borefield, borefield.load)
@@ -1249,8 +1253,7 @@ def test_with_titled_borefield():
     # define params
     ground_data = GroundFluxTemperature(1.9, 10)
     pipe_data = DoubleUTube(1.5, 0.013, 0.016, 0.4, 0.035)
-    fluid_data = FluidData(mfr=0.2)
-    fluid_data.import_fluid_from_pygfunction(gt.media.Fluid('MPG', 30, 2))
+    fluid_data = TemperatureDependentFluidData('MPG', 30).create_constant(2)
     load_data = MonthlyBuildingLoadAbsolute(
         np.array([.176, .174, .141, .1, .045, 0, 0, 0, 0.012, 0.065, 0.123, 0.164]) * 8 * 1350,
         np.array([0, 0, 0, 0, .112, .205, .27, .264, .149, 0, 0, 0]) * 4 * 700,
@@ -1268,17 +1271,13 @@ def test_with_titled_borefield():
     borefield.ground_data = ground_data
     borefield.pipe_data = pipe_data
     borefield.fluid_data = fluid_data
+    borefield.flow_data = ConstantFlowRate(mfr=0.2)
     borefield.set_max_avg_fluid_temperature(17)
 
     assert np.isclose(borefield.depth, 150 * math.cos(math.pi / 7) + 0.75)
     assert np.isclose(borefield.ground_data.calculate_Tg(borefield.depth, borefield.D), 12.157557845032045)
 
     assert np.isclose(borefield.size_L3(), 111.5821049845363)
-
-
-def test_warning_custom_gfunction():
-    with pytest.warns(DeprecationWarning):
-        Borefield(custom_gfunction=CustomGFunction())
 
 
 def test_Rb_and_Re_with_temperture_dep_data():